+++ js/primitivenotes.js
@@ -6,7 +6,7 @@
  * @copyright Copyright (c) 2024, Offerel
  * @license GNU General Public License, version 3
  */
-var mde, tagify;
+var mde, tagify, originalData;
 var filelist = [];
 var loader = document.createElement("div");
 var ldr = document.createElement("div");
@@ -186,22 +186,12 @@ window.rcmail && rcmail.addEventListener("init", function(a) {
 	});
 
 	window.addEventListener("beforeprint", e => {
-		e.preventDefault();
-		e.stopPropagation();
-		var a = window.open('', '', 'height=900, width=900');
-		a.document.write('<html>');
-		a.document.write('<head>');
-		a.document.write('<title>Test</title>');
-		a.document.write('<link rel="stylesheet" type="text/css" href="plugins/primitivenotes/js/highlight/styles/default.css"></link>');
-		a.document.write('<link rel="stylesheet" type="text/css" href="plugins/primitivenotes/skins/print.css"></link>');
-		a.document.write('</head>');
-		a.document.write('<body >');
-		a.document.write(mde.markdown(mde.value()));
-		a.document.write('</body></html>');
-		setTimeout(function() {
-			a.print();
-		}, 100);
-		a.document.close();
+		originalData = document.body.innerHTML;
+		document.body.innerHTML = "<html><head><link rel=\"stylesheet\" type=\"text/css\" href=\"plugins/primitivenotes/skins/print.css\"></link><link rel=\"stylesheet\" type=\"text/css\" href=\"plugins/primitivenotes/js/highlight/styles/default.css\"></link><title>" + document.getElementById('headerTitle').value + "</title></head><body>" + mde.markdown(mde.value()) + "</body>";
+	});
+
+	window.addEventListener("afterprint", e => {
+		document.body.innerHTML = originalData;
 	});
 	
 	document.addEventListener("keyup", event => {
@@ -500,24 +490,6 @@ function downloadNote(note) {
 	xhr.responseType = "blob";
 	xhr.open('GET', location.href + '&_action=getNote&_name=' + note);
 	xhr.send();
-}
-
-function printnote(e) {
-	//e.preventDefault();
-	var a = window.open('', '', 'height=500, width=500');
-	a.document.write('<html>');
-	a.document.write('<head>');
-	a.document.write('<title>Test</title>');
-	a.document.write('<link rel="stylesheet" type="text/css" href="plugins/primitivenotes/js/highlight/styles/default.css"></link>');
-	a.document.write('<link rel="stylesheet" type="text/css" href="plugins/primitivenotes/skins/print.css"></link>');
-	a.document.write('</head>');
-    a.document.write('<body >');
-    a.document.write(mde.markdown(mde.value()));
-    a.document.write('</body></html>');
-	setTimeout(function() {
-		a.print();
-	}, 200);
-	a.document.close();
 }
 
 function sidebyside() {